Implemented Settings State and Key Rebind functionality (Homework for 24/11)
Implemented the settings state.
Added functionality to rebind the keys for the actions defined in the player class.
Keys can be rebound through a UI in the settings state (with a label and a button for each action -> the button can be toggled to listen for user input to rebind).